requirements.txt
Flask
SQLAlchemy
psycopg2
passlib
flask_login
PyJWT